+++ CHANGELOG.md
@@ -43,6 +43,7 @@
 - [5480](https://github.com/vegaprotocol/vega/issues/5480) - Assure indicative trades are in line with actual uncrossing trades
 - [5556](https://github.com/vegaprotocol/vega/issues/5556) - Fix id generation seed
 - [5361](https://github.com/vegaprotocol/vega/issues/5361) - Fix limits for proposals
+- [5557](https://github.com/vegaprotocol/vega/issues/5427) - Fix oracle status at market settlement
 
 ## 0.52.0